Fix clippy lints (rust 1.95)

+++ packages/blitz-paint/src/render/background.rs
@@ -173,8 +173,8 @@ impl ElementCx<'_> {
             BackgroundSizeComputeMode::Size(svg_size.width(), svg_size.height()),
         );
 
-        let x_ratio = (bg_size.width as f64 / svg_size.width() as f64) * self.scale;
-        let y_ratio = (bg_size.height as f64 / svg_size.height() as f64) * self.scale;
+        let x_ratio = (bg_size.width / svg_size.width() as f64) * self.scale;
+        let y_ratio = (bg_size.height / svg_size.height() as f64) * self.scale;
 
         let bg_pos = compute_background_position(
             bg_styles,
